Upgraded to latest version of MALLET (in some cases, the upgrade needs checking - these are marked by TODOs)
Changed how fanseparser is included. Had originally planned on using non-maven-jar, but I couldn't get it working on my Mac so opted for the hackier system scope. It's a bad way of doing it, but it is a lot more robust. Really, the authors of fanseparser should be encouraged to add it to Maven.

Started tidying up the code, however there is a lot of work that needs doing there. There is also huge amounts of dead and redundant code which could be removed - didn't want to do that here, but it might be worth doing in the future.

+++ src/main/java/edu/nyu/jet/hmm/HMMarc.java
@@ -3,7 +3,6 @@ package edu.nyu.jet.hmm;
 
 import java.util.*;
 import java.io.*;
-import edu.nyu.jet.tipster.*;
 
 /**
  *  an arc (from one state to another) in a Hidden Markov Model.
@@ -48,7 +47,7 @@ public class HMMarc {
 
 	void computeProbabilities (int stateCount) {
 		if (stateCount > 0)
-			probability = Math.log(((double) count + 0.01) / (double) stateCount);
+			probability = Math.log((count + 0.01) / stateCount);
 		else
 			probability = HMM.UNLIKELY;
 		if (HMM.probReport)
